requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface
sentence-transformers==3.0.1
qdrant-client==1.12.0
pypdf==4.0.2
python-docx==1.1.0
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.6
passlib[bcrypt]==1.7.4
argon2-cffi
python-jose==3.3.0
python-multipart==0.0.22
sqlalchemy==2.0.48
